+++ packages/contracts/contracts/core/DAO.sol
@@ -14,6 +14,7 @@ import "./component/CallbackHandler.sol";
 import "./permission/PermissionManager.sol";
 import "./IDAO.sol";
 import "./IEIP4824.sol";
+import {hasBit, flipBit} from "../utils/BitMap.sol";
 
 /// @title DAO
 /// @author Aragon Association - 2021
@@ -56,6 +57,9 @@ contract DAO is
     bytes32 public constant REGISTER_STANDARD_CALLBACK_PERMISSION_ID =
         keccak256("REGISTER_STANDARD_CALLBACK_PERMISSION");
 
+    /// @notice Only allows 256 actions to execute per tx.
+    uint256 internal constant MAX_ACTIONS = 256;
+
     /// @notice The [ERC-1271](https://eips.ethereum.org/EIPS/eip-1271) signature validator contract.
     IERC1271 public signatureValidator;
 
@@ -65,8 +69,12 @@ contract DAO is
     /// @notice The [EIP-4824](https://eips.ethereum.org/EIPS/eip-4824) DAO uri.
     string private _daoURI;
 
+    /// @notice Thrown if action length is more than MAX_ACTIONS.
+    error TooManyActions();
+
     /// @notice Thrown if action execution has failed.
-    error ActionFailed();
+    /// @param index Index of action in the array that failed.
+    error ActionFailed(uint256 index);
 
     /// @notice Thrown if the deposit or withdraw amount is zero.
     error ZeroAmount();
@@ -172,21 +180,39 @@ contract DAO is
     }
 
     /// @inheritdoc IDAO
-    function execute(bytes32 callId, Action[] calldata _actions)
+    function execute(
+        bytes32 callId,
+        Action[] calldata _actions,
+        uint256 allowFailureMap
+    )
         external
         override
         auth(address(this), EXECUTE_PERMISSION_ID)
-        returns (bytes[] memory)
+        returns (bytes[] memory execResults, uint256 failureMap)
     {
-        bytes[] memory execResults = new bytes[](_actions.length);
+        if (_actions.length > MAX_ACTIONS) {
+            revert TooManyActions();
+        }
 
-        for (uint256 i; i < _actions.length; ) {
-            (bool success, bytes memory response) = _actions[i].to.call{value: _actions[i].value}(
+        execResults = new bytes[](_actions.length);
+
+        for (uint256 i = 0; i < _actions.length; ) {
+            address to = _actions[i].to;
+            (bool success, bytes memory response) = to.call{value: _actions[i].value}(
                 _actions[i].data
             );
 
-            if (!success) revert ActionFailed();
+            if(!success) {
+                // If the call failed and wasn't allowed in allowFailureMap, revert.
+                if(!hasBit(allowFailureMap, uint8(i))) {
+                    revert ActionFailed(i);
+                }
 
+                // If the call failed, but was allowed in allowFailureMap, store that 
+                // this specific action has actually failed.
+                failureMap = flipBit(failureMap, uint8(i));
+            }
+            
             execResults[i] = response;
 
             unchecked {
@@ -194,9 +220,7 @@ contract DAO is
             }
         }
 
-        emit Executed(msg.sender, callId, _actions, execResults);
-
-        return execResults;
+        emit Executed(msg.sender, callId, _actions, failureMap, execResults);
     }
 
     /// @inheritdoc IDAO