+++ packages/contracts/test/voting/multisig.ts
@@ -494,8 +494,7 @@ describe('Multisig', function () {
         );
         expect(proposal.parameters.startDate).to.equal(startDate);
         expect(proposal.parameters.endDate).to.equal(endDate);
-        expect(proposal.tally.approvals).to.equal(0);
-        expect(proposal.tally.addresslistLength).to.equal(1);
+        expect(proposal.approvals).to.equal(0);
         expect(proposal.actions.length).to.equal(0);
 
         expect(await multisig.canApprove(id, signers[0].address)).to.be.true;
@@ -533,7 +532,7 @@ describe('Multisig', function () {
         );
         expect(proposal.parameters.startDate).to.equal(startDate);
         expect(proposal.parameters.endDate).to.equal(endDate);
-        expect(proposal.tally.approvals).to.equal(1);
+        expect(proposal.approvals).to.equal(1);
       });
 
       it('increases the proposal count', async () => {
@@ -567,7 +566,7 @@ describe('Multisig', function () {
       await expect(
         multisig.createProposal(dummyMetadata, dummyActions, true, false, 5, 0)
       )
-        .to.be.revertedWithCustomError(multisig, 'InvalidStartDate')
+        .to.be.revertedWithCustomError(multisig, 'DateOutOfBounds')
         .withArgs(timeStamp, 5);
     });
 
@@ -577,7 +576,7 @@ describe('Multisig', function () {
       await expect(
         multisig.createProposal(dummyMetadata, dummyActions, true, false, 0, 5)
       )
-        .to.be.revertedWithCustomError(multisig, 'InvalidEndDate')
+        .to.be.revertedWithCustomError(multisig, 'DateOutOfBounds')
         .withArgs(timeStamp, 5);
     });
   });
@@ -679,7 +678,7 @@ describe('Multisig', function () {
       });
 
       it('approves with the msg.sender address', async () => {
-        expect((await multisig.getProposal(id)).tally.approvals).to.equal(0);
+        expect((await multisig.getProposal(id)).approvals).to.equal(0);
 
         let tx = await multisig.connect(signers[0]).approve(id, false);
 
@@ -688,7 +687,7 @@ describe('Multisig', function () {
         expect(event.args.approver).to.not.eq(multisig.address);
         expect(event.args.approver).to.eq(signers[0].address);
 
-        expect((await multisig.getProposal(id)).tally.approvals).to.equal(1);
+        expect((await multisig.getProposal(id)).approvals).to.equal(1);
       });
 
       it("reverts if the proposal hasn't started yet", async () => {